Change create/delete page to alter order
Pages form a linked list. Currently the client is responsible for
ensuring the consistency of the list.

This commit changes this by making sure new pages are 'appended' to the end
of the list. Deleting a page in the middle of the page list now connects
the two lists together. These changes mean there should only be one
connected list of pages.

This means clients are no longer able to create pages which random
'next' values. The update route hasn't been changed to keep that
possibility available, but maybe should be closed off too to ensure
consistency.