Add /ping endpoint for forms-api
This endpoint will be used to check whether the service is online.

This endpoint is public and to help avoid having a `api/v1/../ping`
I decided to breakout all the api endpoints into a seperate class
and then to mount that class to the server. Apart from making the
url /ping a bit clearer, it also meant that the endpoint wouldn't
require the api key like the rest of the api endpoints.

Can view it by visiting http://localhost:9292/ping