Switch is_optional from TrueClass to String
None of our other params have specific types. There also had a problem
with the requests coming from forms-admin where the value was being sent
as strings in the call. The database column is setup as boolean so should
fail any requests that try to set it to non-boolean values.