Move most of original ruby-grape app to new folder
We plan to move away from ruby-grape to rails-api purely because most of
the team aren't familiar with ruby and its making it hard for all devs
to contribute to forms service when the service has a combination of
two very different frameworks and conventions.

We agreed that we would continue to use 3 seperate microservices
(forms-admin, forms-api & forms-runner) but this will hopefully reduce
the need for us to make it monolith. We will monitor the situation and
if we do still need to merge all the microservices we would be
in a better position as all the services would be using rails conventions