Add rspec request tests
Most of the current tests are unit tests and don't cover the controller.
There are tests for the repositories, migrations and page domain object
as well as some tests which cover the API_KEY and ping routes.

This commit adds request tests for the forms and pages routes.

The tests aim to cover the external API so that large changes can be
made while preserving the interface and calling conventions.