Fix FK violation error when deleting forms that have pages
Rails was not deleting the pages that existed for a form that was
about to be deleted. This happened locally on my machine and on
staging when testing out the rails api version.

Error message:

PG::ForeignKeyViolation: ERROR:  update or delete on table "forms" violates foreign key constraint "fk_rails_ffa7b626f0" on table "pages"
DETAIL:  Key (id)=(256) is still referenced from table "pages".

Co-authored-by: Samuel Culley <25597009+SamJamCul@users.noreply.github.com>