created make_live! model method
this is being used in the controller and the idea
is that we will eventually update this implementation
to create a published form either using paper trail, our own published forms table or some other means. The benefit is this we do not have
to touch the controllers again and can concentrate on
the models methods that return what we need.

Co-authored-by: Tom Iles <tom.iles@engineering.digital.dwp.gov.uk>