Allow live forms to be deleted
This is a temporary fix for the time being. We have a ticket
to look at deleting forms in general and we will need to think
how we delete forms in general. Also from an e2e POV, we go
through the whole process of creating a form, making it live,
filling it in and then after a form submission we try to delete
the live form. Without this temporary fix our e2e tests will continue
to fail when trying to delete the live form.