Refactored access_token.rb model
We decided that instead of using bcrypt to hash our tokens we would
instead use sha-256. Although sha-256 is not a bulletproof as bcrypt it
allows use to regenerate the hash without having to use a salt. bcrypt
by default uses a dynamic generated salt and so when it hashes the token
everytime it would use a different salt. This makes it difficult to search
the table for a particular record without iterating through each record,
using bcrypt to compare the hashed db value with the users token.