Make sure datetimes are consistent when making a form live
It is important when we make a form live that we are able to reliably
pinpoint the time when we made a form live and compare that time to the
time of other events. This commit makes sure that the time when we make
a form live is aligned with the time when the form was last updated and
the time when the made_live_form entry is created.

Note that this was supposed to be the case from PR #207, but it was
unfortunately missed.