Move logic to update form.question_section_completed out of controller
We want to be able to write model specs that update pages, while making
sure that the logic surrounding form.question_section_completed is
updated in the same way as would be done if the controller was used.
This PR moves the logic out of the controller and into the model.

Note that although we could use callbacks to achieve the same behaviour,
we want to be explicit, so that we don't accidentally affect what form
creators see when we do a migration or editing via the rails console.