Fix form-runner
Commit 38b3a2c broke the forms-runner for newly made live forms, as form
blob did not include the live_at attribute required for validation of
the form [[1]], [[2]]. This manifested as `invalid live_at times` [[3]].

For backwards-compatibilty, this commit adds the live_at datetime back
into the form blob for made live forms. For draft form versions, live_at
is not needed, so we leave it out.

[1]: https://github.com/alphagov/forms-runner/blob/main/app/models/form.rb#L25-L30
[2]: https://github.com/alphagov/forms-runner/blob/31a0d62dab8a950d0df59e9a652c5006d9362d63/app/controllers/forms/base_controller.rb#L66
[3]: https://govuk-forms.sentry.io/issues/4034044478/