Include conditions in json bundling api calls
Any conditions that are associated with a page via the "routing_conditions" associtation will now be included with the page.

The result is a new key being available on pages, `routing_conditions`, which will contain an array of all the page's associated routing conditions, or an empty array if there aren't any.