Standardise 400/404 handling
It seems pointless to have to define a 400/404 exception handling for each
new controller we create.  This just moves the functionality to
the application controller where all controllers can use it.

I also have to update one test to pass in the param that the url need to
return 404 error rather than a 400.