Add validation_errors method to Condition model
We want to dynamically check if a route condition is valid. In this
commit, we create a `validation_errors` method which will be added
the the json response. The method will return an array of error objects
for the admin to display to the user. in the array there will be a call
to each validation check method that we build.