Fix live_at Time
Specifies the order for made_live_forms so that the ordering and
selection of the latest version with ActiveRecord.last function
correctly selects the latest version based on the created_at time.
Without this the ordering is inconsistent because the resulting `SELECT`
statement does not include an `ORDER BY` clause; postgres will varying
the order of returned rows without an `ORDER BY` clause due to the
`synchronize_seqscans` optimisation.

Updates the integration tests to create multiple live versions and
ensure the returned `live_at` is based upon the create_at time of the
latest version.