Align Github Workflow Across Apps
This commit aligns the ruby on rails workflow with out other apps. The
matrix testing of Ruby versions was not being used since it contained a
single version and it is simpler to use the standard ruby/setup-ruby
workflow which will install the version of ruby in the .ruby-version
file.