Delete existing routes if answer_type is not selection
We warn users when they edit their pages and changing
the answer type will delete any existing routes.

While reviewing this part of the page, it was noted
that the current implementation would reset the
forms "question section completed" attribute everytime
a user pressed "save" regardless as to whether there was anything to be changed. This has been changed in this commit
so that we only attempt to save the page and modify form attributes
if there are actually changes.