Delete conditions when changing selection settings
We want to ensure that conditions that check against selection questions only exist so long as only one option can be selected. It's currently not possible to set up conditions against selections that can have multiple options, so if that change is made, conditions relying on the page will be destroyed.

This also refactors deleting conditions when the answer type changes, so that the check_conditions are removed. This is currently identical to removing the routing_conditions, but in the future we may differentiate between them.