Add URL query params to logging messages
When a HTTP request includes query parameters it is useful to be able to
see these in our logs. By default Lograge filters this out as it can
contain sensitive information [[1]], but we don't have anything like
that in our system, so we add in the params hash following the
suggestion in the Lograge readme [[2]]. Splunk will then just do the
right thing with the nested JSON to make it easy to work with (thanks
@danworth for pointing this out!).

This is especially useful at the moment when we are changing from using
`org` slug to `organisation_id` (see PR #295), as without this logging
we can't see what the admin app is sending to the api app for requests
to `/forms`.

[1]: https://github.com/roidrage/lograge/blob/09bd96e1cb56f94841a64473e4195d413ce908fc/lib/lograge/log_subscribers/action_controller.rb#L21-L39
[2]: https://github.com/roidrage/lograge/tree/v0.12.0#what-it-doesnt-do