Change CRUD API endpoints to respond with resource
We want the API to respond with a resource when requests are made, even
if it's a request to update or delete an object. As well as being more
in line with other APIs (see the JSON::API spec for example [[1]]), this
will also let us reduce the number of requests we need to make in the
admin app.

This commit just deals with our simple traditional Create Read Update
and Delete operations, which have an obvious response. Technically this
is a breaking change and should result in a new API version, but
currently our API is internal only and none of the consumers rely on the
response body data, so no breakage has occured.

For the destroy actions we can just respond with 204 No Content, this is
one of the responses considered valid in the JSON::API spec [[2]].

[1]: https://jsonapi.org/format/
[2]: https://jsonapi.org/format/#crud-deleting-responses