Change response of make live endpoint to respond with resource
We want our API endpoints to respond with resources where possible,
instead of the success JSON payload, which isn't really useful.

For the make form live endpoint, it makes most sense to respond with the
JSON of the new live form object. This doesn't really fit with the
semantics of a RESTful API, but is more useful than what we do now and
still compatible with our existing API consumers.