Fix logging of form_id in FormsController
For requests that are routed to the FormsController, the form ID is
named `:id` rather than `:form_id`, so the logging setup needs to be
slightly different. This commit adds a specialisation of the
`append_info_to_payload` method to the `FormsController` class so we get
the correct behaviour when necessary.

Thanks @tomisles for pointing out we can use OOP to fix this issue.