Add selection question to database seed
Adds a question asking for an answer from a list of options to the test
form in the database seed.

While adding this change I learnt that we are setting the value of
"only_one_option" to "0" if the user does not check the "People can only
select one option" when creating their question [[1]]. This seems weird
to me, but isn't a bug per-se? I've added a note to investigate this in
the future.

[1]: https://github.com/alphagov/forms-admin/blob/7ff8e352e0c237878781a7b91bdb89b7e777c3dd/app/views/pages/selections_settings.html.erb#L27-L28