Add task to convert what_happens_next to markdown
Adds a rake task which uses the ReverseMarkdown gem to convert each form
and made_live_form's what_happens_next_text content to markdown, and
save this markdown to the what_happens_next_markdown field. Also adds a
task to clear the content in the what_happens_next_markdown field in
case we need to revert the change.