Rename authentication_key to auth_key
In other apps we call this setting `Settings.forms_api.auth_key`. It
would be helpful if this setting had the same name in all apps, because
we could set it everywhere with one environment variable.

This commit renames the setting key in forms-api to match what is
currently in forms-admin and forms-runner.