example of how archive may work
by calling the make-unlive endpoint the form state is transitioned to :archived. When we request a live version if the form is archived then we return 404 instead of the last made live version.

The issue with this implementation is that it also means that when forms-admin requests `/live` endpoint to show read only version of the live form, it too will return a 404 we might have to try identify if the request host is forms-admin or forms-runner and ignore the archived condition to return last live version but the form would still have its state we should display to the user.