add FormStateMachine with state transitions
This is a basic state machine for forms which mimics the existing behaviour we have within the service. This state machine has not been added to the forms model yet.

When a form is in draft state, it can be HARD deleted because it is "Work in progress"/unofficial form in this state. This initial state can transition to `live` state once all the required sections for a form have been completed.

When a form is made live, we take a snapshot of the form record along with pages and conditions and store it in a table called `made_live_forms`. A form can have multiple made_live_forms records for each time it has been made live.

In this state, a form cannot be HARD deleted because it is now an official document and it could have been used by the general public to submit data. In future, we will have an archive feature to stop people being about to fill in live forms.

When a form is in Live with draft, it means that the forms has been made live at least once but there are some changes to the form record which have not been made live yet. Currently users cannot `revert`/delete changed that made since the form was last live. The work around for this is for them to manually update the draft until its in good state that the user is happy to "make live" again.

```mermaid
stateDiagram-v2
    Draft --> Live : make_live!
    Draft --> Deleted : delete_form!
    live_with_draft: Live with draft
    Live --> live_with_draft : create_draft_from_live_form!
    live_with_draft --> Live : make_live!

```