exclude state from stored snapshots
This issue was identified by a failing spec which wasn't expecting the state column. We should exclude the state from snapshots and look to the form object itself for the latest state. If we need to include it as part of the response we could dynamically add it to the snapshot before returning the json data.