tidy up migrations into a single migration file
it good to periodically clean up migration files because it can be difficult to rebuild the db from scratch when you start implementation when required columns etc. We also had a mix of data migration and schema migrations. I reviewed the data migrations and none of them look that important that we would need to keep them as the existing data has already run these. They were also created in a time when we didn't have a way to run rake tasks in AWS.

We already had one issue with timestamps which was fixed here https://github.com/alphagov/forms-api/commit/28b81d344f668cb922054a216e60bdcb023bdccc and while trying to add statemachine to forms model it was picked up in the spike and the final implementation that data migrations were causing and issue because it was expecting a state column in the data migration which hadn't yet been created but was declared in Rails model

https://github.com/alphagov/forms-api/pull/414
https://github.com/alphagov/forms-api/pull/425

I followed this guide https://naturaily.com/blog/remove-old-migrations-ruby-on-rails.