set the existing forms states for state machine
THIS IS A ONE TIME MIGRATION AND COULD BE REMOVED ONCE RUN IN PRODUCTION

CAUTION was used while updating the state column to not update the `updated_at` column which would put live forms into live_with_draft state incorrectly.

we already have a has_live_version and has_draft_version which we will use to determine the current state of the form. Based off a combination of the two methods we can set the correct state for the form.

once this has been implemented we will do another pr to refactor these two methods to use the state machine states instead.