remove page_order column from forms table
This column is obsolete and was created before Jan 2023 when forms-api was not using rails. It was initial created to start an array of page ids in a particular order but when we moved over to rails we switched over to using acts_as_a_list gem which handled positing and ordering of associated records. I checked in forms-admin adn forms- runner and none of the code uses it nor does any newish form include it.