Set make_live_status for archived form
If a form has state `archived`, make_live_status had a value of nil.
This caused an exception to be thrown in forms-admin when loading the
page to create a draft version of the form.

Default this method to return "completed" if the form doesn't have a
draft version - as this implies that the form must be in either `live`
or `archived` state.