Use Page::ANSWER_TYPES constant to tidy service
Our report was previously using a manually maintained list of answer
types. This is fragile and increases the maintenance burden when adding
or removing answer types in future. This commit updates the service to
use the existing Page::ANSWER_TYPES constant, so the list of answer
types only needs to be updated in one place.

Also updates the request test for this controller action to stop it
breaking when the order of keys in the response body changes.