Sets external id for new forms
In order to start using the external id as the replacement for the id, we'll start off by doing an invisible shift behind the scenes. This means that intially, we'll want the external id and id to be the same value, so that we can transition over. We still need to migrate the old forms so that they also have their external id set.