Add service object for getting v2 API form documents
We want to be able to get v2 form documents for v1 forms, and in future
get v2 form documents from a table.

It would be nice if there was one unified interface for doing this. I
think it makes sense to use something like the repository pattern, where
we have a class that gets the form document from the right place.

Putting this in a service object allows us to keep the logic out of our
models.

I've tried to make the methods similar to methods that might be on a
model so in a future where we only have one way of getting form
documents we can replace the service object with a normal simple model.