Tell ActiveRecord to ignore organisation_id column
Adding the `organisation_id` to the ignored_columns for the Forms
ActiveRecord ensures there is no code that raises errors due to having
a cached schema in memory at the time the schema migration is run.

Once this is deployed, we can safely merge the migration to drop the
column from the table.

We can also change an assertion in `forms_controller_spec` back to
`match` rather than `include` now that the `organisation_id` column is
ignored by ActiveRecord.