Delete redundant migration
The db/migrate/20240126213348_set_existing_form_state.rb migration used
the Form model. This meant that if the Form model was updated with new
validations, then this migration would fail as the Form would be
invalid.

The migration itself is exclusively a data migration which is no longer
needed, so can be deleted.

Co-authored-by: Stephen Daly <stephen.daly@digital.cabinet-office.gov.uk>