Add api_v2_form_documents table and model
To represent `form_documents` which will be used to model the v2 API we
need a new model and database table.

This model will hold the 'content' of a form as a JSON structure. A form
can have more than one form_document and will be referenced by its tag.

Tags will be used to represent the different states of a form. For
example most forms will have a form_document with a tag draft. Forms
which have been made live will have a form with the tag 'live' and
archived forms will have a form document with the tag 'archived'.

Currently there is no validation on the form_document, other than it has to be
present. There is also no index on tag, which might be something to add
in the future if needed.

FormDocuments belong to Api_v2_forms rather than the standard Form.

This commit doesn't include the association on Api_v2_forms. That should
be added as needed.