Update form models to allow deleting
The table api_v2_form_documents uses form_id as a foreign key. Deleting
a form will raise an error if there are associated form_documents.

This commit adds a has_many association to the Form and
Api::V2::Form models. This ensures that any form_documents are deleting
before any Form is removed to that no error is thrown from the db.