Add Api::V2::ModelSync class
Uses of the V1 API should update the V2 FormDocument so they can be
used in the responses from the V2 API.

This commit introduces a new class, awkwardly name ModelSync. This class
will be used in the V1 API to keep FormDocuments in sync.

It calls the conversion function with the contents of a form_blob and
creates, updates and deletes form_documents as needed.