Allow getting draft form from v2 API even when form does not have draft version
This commit changes the v2 form document repository so that a draft form
document can always be retrieved, even if the v1 form returns false to
`#has_draft_version`.

This maintains backwards compatibility of v2 API with v1 API and fixes
preview links in forms-admin when forms-runner is using v2 API.

Ideally we would in future consider whether this behaviour is what we
want, or if the logic can be moved to forms-runner or forms-admin
somehow.