Add share_preview_completed column to forms table
We are adding a new task to the draft form task list to share a
preview of the form. This task will have radio buttons to mark the
task as completed.

Add migration to add a new `share_preview_completed` boolean column to
the forms table to store whether the new task has been completed.

Default the value to `false`. This means that if a draft version of an
existing live form is created, this task will need to be completed
before the draft can be made live.