Add task status for sharing a preivew of the form
Check the value of the `share_preview_completed` flag on the
form to determine the status of this new task.

This will be a mandatory task, but for now don't add a check to the
`incomplete_tasks` method as we need to update the pages for the new
task in forms-admin first.