Add new submission type of s3
When the submission type is set to s3 we will send the submission data
to an s3 bucket.

We have also added a rake task in order to set the submission type to
s3. In order to be able to switch back to an email submission type we
also added a rake task to set the submission type to email.

Co-authored-by: Stephen Daly <stephen.daly@digital.cabinet-office.gov.uk>