Update FormDocuments when changing conditions
When the conditions controller as part of V1 API is called to create,
update or destroy a condition, the form document should be updated to
reflect the change.

This commit adds calls to the controller to the FormDocumentSyncService
to ensure the form document is updated.

The Condition model has methods which save and update the form. It's an
option to extend these methods to update the form document.

This would make these mehods more complicated, so the calls are put in
the controller to match the calls in the page controller.