Add s3_bucket_region column to forms
This column will store the region the S3 bucket was created in. When
using the AWS SDK for Ruby, it is required to provide the bucket
region. If the region is different to the region the bucket was
created in, we get the following error:

Aws::s3::Errors::PermanentRedirect - The bucket you are attempting to
access must be addressed using the specified endpoint. Please send all
future requests to this endpoint.