Update rake task to set S3 bucket region
Update the rake task that sets the submission_type for a form to "s3"
to also set the s3_bucket_region to the provided value.

We need to be careful about which regions we are sending data to.
We would only expect government departments to create buckets in
eu-west-1 (Ireland) or eu-west-2 (London). If we allow any other
region we might run into information assurance issues with where the
data is stored. Restrict the rake task so that only these 2 regions
are permitted.