Allow answer value to be nil for condition reached after another condition
We want to model a branching route in a form as two conditions with the
same check page but different routing pages: the condition for the first
page reached skips the pages in the first branch if it is met, and the
condition for the page at the end of the first branch skips the pages in
the second branch. Because the second condition can only be reached in
the case where it should skip, it is redundant to specify the answer
value, so we want to be able to omit it.

This commit changes the `warning_answer_doesnt_exist` validation so that
if a condition is after another condition then `answer_value` is allowed
to be `nil`.