Use routing_page_position in condition validations
Validation to check that a condition is not routing to the
next page, making it redundant, uses the check_page.position.

This doesn't work for secondary skips. Secondary skips use the
check_page attribute to show which route they belong to but route from
the routing_page.

This commit changes the attribute used to check position in
warning_routing_to_next_page. This means that errors will be reported
for secondary skips as a well as normal conditions.

Existing conditions, where routing_page_id and check_page_id should be
the same, are not affected.