Use routing_page_position in goto_page check
Validation to check that a condition is not routing to the
a page before the route use the check_page attribute.

This doesn't work for secondary skips. Secondary skips use the
check_page attribute to show which route they belong to but route from
the routing_page.

This allows form creators to make forms with loops live.

This commit changes the attribute used to check position in
warning_goto_page_before_check_page. This means that errors will be
reported for secondary skips as a well as normal conditions.

Existing conditions, where routing_page_id and check_page_id should be
the same, are not affected.