Refactor #warning_goto_page_before_check_page
Refactor the validation method for conditions.

The method no longer relies on check_page so we don't need to check for
nil.

We also can also change the main test to:

goto_page.position <= routing_page.position

This means the error will be present on any question which goto's itself
or an earlier question.

We don't allow conditions to be created with a self referencing goto
value but it was a case which the previous code did not cover.