Fix controller specs
The request specs for the Condition controller no longer pass.

This is a result of the change to the validation on Condition which
checks the goto page is valid.

This happens because the change to the condition validation code means
the setup used in the request tests create a condition with a validation
error.

To test for equality the test converts the condition into json, then
parses it back to a ruby object and calls symbolize_keys before
comparing against the result of the call.

Symbolize_keys only symbolizes the top level keys, so the values in the
validation_errors array don't get symbolized and so the equality test
fails.

One option would be to change the test setup to create valid conditions.
Conditions can have errors though so I haven't change them.

Instead, the equality check has been changed to
from:

```
expect(json_body[i]).to eq(JSON.parse(p.to_json).symbolize_keys)
```

to:

```
expect(json_body[i]).to match(p.as_json.deep_symbolize_keys)
```

I think this is probably a clearer way to asset the two hashes are
equal.