Fix page model specs
The page model specs fail when testing the #has_routing_errors method.

The tests check the method when a page's condition has returns either
true or false for the validation_error.

They fail because the tests previously relied on creating a Condition
without check_id. This would previously not result in a validation
error.

Now the validations don't require check_id to be set, the tests need to
be adapted to account for this.

I looked at fixing the setup to make the condition valid and invalid.

This resulted in more test setup than felt right to ensure the
validations pass. To correctly set up, a form needs to be created with
the right pages and a correct condition created.

Page#has_routing_errors is a simple method which gets most of it's
behaviour from calling has_routing_errors on condition.

Instead of replicating the validation test setup for conditions here,
I've stubbed the condition to return the values we need to test the
method.

This requires less setup, is quicker to run and keeps the test self
contained. This comes at the cost of no longer being an integration
test.