Change page model to delete secondary skip conditions when skip condition is deleted
As part of deleting all routes for a page when it is deleted, we need to
delete any routes that depend on routes to be deleted.

This is not a relationship that is expressed in our model directly,
but it does exist in the case of a page with a skip condition and a
secondary skip condition.

Because there isn't a recorded association, we need to jump through some
hoops to make sure everything is destroyed the way we want.