Add report to get all live forms
Add API endpoint to retrieve all live forms for the purpose of
generating reports. This uses the "api-pagination" and "pagy" gems to
paginate the results to avoid sending all the form data in one API
response.

We return all the form data for forms-admin to process so that
organisation and group data can be added to the report.